requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.6.3
blinker==1.4
boto==2.44.0
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
cloud-init==18.3
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
distro==1.0.1
distro-info==0.18
et-xmlfile==1.0.1
google-compute-engine==2.8.2
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
jdcal==1.4
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lxml==4.2.4
MarkupSafe==1.0
numpy==1.15.0
oauthlib==2.0.6
openpyxl==2.5.5
PAM==0.4.2
pandas==0.23.4
parsel==1.5.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
PyDispatcher==2.0.5
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.2
python-dateutil==2.7.3
python-debian==0.1.32
pytz==2018.5
pyxdg==0.25
PyYAML==3.12
queuelib==1.5.0
requests==2.18.4
requests-unixsocket==0.1.5
Scrapy==1.5.1
scrapy-rotating-proxies==0.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
ssh-import-id==5.7
systemd-python==234
Twisted==17.9.0
typing==3.6.4
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
w3lib==1.19.0
xlrd==1.1.0
XlsxWriter==1.0.5
zope.interface==4.3.2